all files pushed

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,15 +8,16 @@
  * Return: the number of elements in a linked list
  */
 
-int list_len(const list_t *h)
+size_t list_len(const list_t *h)
 {
-	int i;
-	
+	size_t i;
+
 	i = 0;
+
 	while (h != NULL)
 	{
-		 i++;
-		 h = h->next;
+		i++;
+		h = h->next;
 	}
